Refactor: View 그룹 내 파일들을 Model, ViewModel로 역할별 분리
